requirements.txt
git+https://github.com/Rapptz/discord.py
beautifulsoup4
requests
firebase_admin
google-api-python-client 
google-auth-oauthlib
google-auth-httplib2
